# contract/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "BAHI",
    "description": "The udhaar book both sides can see. All data is synthetic.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/shops/{shop_id}/book": {
      "get": {
        "tags": [
          "shopkeeper"
        ],
        "summary": "Get Book",
        "description": "A1. Every figure here is computed by bahi/domain from the ledger.",
        "operationId": "get_book_shops__shop_id__book_get",
        "parameters": [
          {
            "name": "shop_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Shop Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShopBookOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/shops/{shop_id}/counter": {
      "get": {
        "tags": [
          "shopkeeper"
        ],
        "summary": "Get Counter",
        "description": "A2. Everyone who scanned the udhaar QR in the last three minutes.",
        "operationId": "get_counter_shops__shop_id__counter_get",
        "parameters": [
          {
            "name": "shop_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Shop Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CounterOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/shops/{shop_id}/customers": {
      "get": {
        "tags": [
          "shopkeeper"
        ],
        "summary": "List Customers",
        "description": "Everyone in the book, for picking someone who is not at the counter.",
        "operationId": "list_customers_shops__shop_id__customers_get",
        "parameters": [
          {
            "name": "shop_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Shop Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CustomerOut"
                  },
                  "title": "Response List Customers Shops  Shop Id  Customers Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "shopkeeper"
        ],
        "summary": "Add By Name",
        "description": "A4. No phone: kept by name only, like the notebook.",
        "operationId": "add_by_name_shops__shop_id__customers_post",
        "parameters": [
          {
            "name": "shop_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Shop Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NameOnlyIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/shops/{shop_id}/entries": {
      "post": {
        "tags": [
          "shopkeeper"
        ],
        "summary": "Record Entry",
        "description": "Record udhaar: for a scan at the counter, or a customer from the book.",
        "operationId": "record_entry_shops__shop_id__entries_post",
        "parameters": [
          {
            "name": "shop_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Shop Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecordIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntryOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/shops/{shop_id}/accounts": {
      "get": {
        "tags": [
          "shopkeeper"
        ],
        "summary": "Find Account",
        "description": "A4. The Paytm account behind a mobile number or UPI ID, and whether he is\nalready in this book. The number is only used to look, never kept.",
        "operationId": "find_account_shops__shop_id__accounts_get",
        "parameters": [
          {
            "name": "shop_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Shop Id"
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Q"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/shops/{shop_id}/customers/invite": {
      "post": {
        "tags": [
          "shopkeeper"
        ],
        "summary": "Invite",
        "description": "A4. Send invite: he accepts on his own phone, and nothing is recorded\nagainst him until he does.",
        "operationId": "invite_shops__shop_id__customers_invite_post",
        "parameters": [
          {
            "name": "shop_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Shop Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InviteIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AccountOut": {
        "properties": {
          "person_id": {
            "type": "string",
            "title": "Person Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "here": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "linked",
                  "invited",
                  "name_only"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Here"
          }
        },
        "type": "object",
        "required": [
          "person_id",
          "name",
          "here"
        ],
        "title": "AccountOut",
        "description": "The Paytm account behind a number or UPI ID. The number is not kept."
      },
...
      "BookOut": {
        "properties": {
          "customer_count": {
            "type": "integer",
            "title": "Customer Count"
          },
          "invited_count": {
            "type": "integer",
            "title": "Invited Count"
          },
          "owing_count": {
            "type": "integer",
            "title": "Owing Count"
          },
          "outstanding_paise": {
            "type": "integer",
            "title": "Outstanding Paise"
          },
          "lines": {
            "items": {
              "$ref": "#/components/schemas/LineOut"
            },
            "type": "array",
            "title": "Lines"
          }
        },
        "type": "object",
        "required": [
          "customer_count",
          "invited_count",
          "owing_count",
          "outstanding_paise",
          "lines"
        ],
        "title": "BookOut"
      },
...
      "CounterOut": {
        "properties": {
          "shop_id": {
            "type": "string",
            "title": "Shop Id"
          },
          "waiting": {
            "items": {
              "$ref": "#/components/schemas/WaitingOut"
            },
            "type": "array",
            "title": "Waiting"
          }
        },
        "type": "object",
        "required": [
          "shop_id",
          "waiting"
        ],
        "title": "CounterOut",
        "description": "Everyone at the counter, oldest scan first."
      },
      "CustomerOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "display_name": {
            "type": "string",
            "title": "Display Name"
          },
          "tag": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tag"
          },
          "joined": {
            "type": "string",
            "enum": [
              "linked",
              "invited",
              "name_only"
            ],
            "title": "Joined"
          }
        },
        "type": "object",
        "required": [
          "id",
          "display_name",
          "tag",
          "joined"
        ],
        "title": "CustomerOut"
      },
...
      "EntryOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "customer_id": {
            "type": "string",
            "title": "Customer Id"
          },
          "display_name": {
            "type": "string",
            "title": "Display Name"
          },
          "shop_id": {
            "type": "string",
            "title": "Shop Id"
          },
          "shop_name": {
            "type": "string",
            "title": "Shop Name"
          },
          "amount_paise": {
            "type": "integer",
            "title": "Amount Paise"
          },
          "paid_paise": {
            "type": "integer",
            "title": "Paid Paise"
          },
          "status": {
            "type": "string",
            "enum": [
              "recorded",
              "confirmed",
              "disputed",
              "corrected",
              "settled"
            ],
            "title": "Status"
          },
          "recorded_at": {
            "type": "string",
            "format": "date-time",
            "title": "Recorded At"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "spoken_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Spoken Text"
          },
          "corrects_entry_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Corrects Entry Id"
          },
          "acknowledged_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Acknowledged At"
          },
          "button": {
            "type": "string",
            "title": "Button"
          }
        },
        "type": "object",
        "required": [
          "id",
          "customer_id",
          "display_name",
          "shop_id",
          "shop_name",
          "amount_paise",
          "paid_paise",
          "status",
          "recorded_at",
          "note",
          "spoken_text",
          "corrects_entry_id",
          "acknowledged_at",
          "button"
        ],
        "title": "EntryOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "InviteIn": {
        "properties": {
          "query": {
            "type": "string",
            "maxLength": 60,
            "minLength": 3,
            "title": "Query"
          },
          "tag": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 40
              },
              {
                "type": "null"
              }
            ],
            "title": "Tag"
          }
        },
        "type": "object",
        "required": [
          "query"
        ],
        "title": "InviteIn"
      },
...
      "LineOut": {
        "properties": {
          "customer_id": {
            "type": "string",
            "title": "Customer Id"
          },
          "display_name": {
            "type": "string",
            "title": "Display Name"
          },
          "tag": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tag"
          },
          "joined": {
            "type": "string",
            "enum": [
              "linked",
              "invited",
              "name_only"
            ],
            "title": "Joined"
          },
          "balance_paise": {
            "type": "integer",
            "title": "Balance Paise"
          },
          "day": {
            "type": "integer",
            "title": "Day"
          },
          "chip": {
            "type": "string",
            "enum": [
              "on_rhythm",
              "changed",
              "not_confirmed",
              "new"
            ],
            "title": "Chip"
          },
          "rhythm": {
            "$ref": "#/components/schemas/RhythmOut"
          }
        },
        "type": "object",
        "required": [
          "customer_id",
          "display_name",
          "tag",
          "joined",
          "balance_paise",
          "day",
          "chip",
          "rhythm"
        ],
        "title": "LineOut"
      },
...
      "NameOnlyIn": {
        "properties": {
          "display_name": {
            "type": "string",
            "maxLength": 40,
            "minLength": 1,
            "title": "Display Name"
          },
          "tag": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 40
              },
              {
                "type": "null"
              }
            ],
            "title": "Tag"
          }
        },
        "type": "object",
        "required": [
          "display_name"
        ],
        "title": "NameOnlyIn"
      },
...
      "RecordIn": {
        "properties": {
          "amount_paise": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Amount Paise"
          },
          "scan_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scan Id"
          },
          "customer_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Customer Id"
          },
          "note": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 80
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "spoken_text": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Spoken Text"
          }
        },
        "type": "object",
        "required": [
          "amount_paise"
        ],
        "title": "RecordIn",
        "description": "An udhaar entry. Either a scan (he is at the counter) or a customer."
      },
...
      "RhythmOut": {
        "properties": {
          "n": {
            "type": "integer",
            "title": "N"
          },
          "median_gap": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Median Gap"
          },
          "max_gap": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Gap"
          },
          "last_paid": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Paid"
          }
        },
        "type": "object",
        "required": [
          "n",
          "median_gap",
          "max_gap",
          "last_paid"
        ],
        "title": "RhythmOut"
      },
...
      "ShopBookOut": {
        "properties": {
          "synthetic": {
            "type": "boolean",
            "const": true,
            "title": "Synthetic",
            "default": true
          },
          "about": {
            "type": "string",
            "title": "About",
            "default": "Synthetic demo data for one seeded shop."
          },
          "today": {
            "type": "string",
            "format": "date",
            "title": "Today"
          },
          "shop": {
            "$ref": "#/components/schemas/ShopOut"
          },
          "book": {
            "$ref": "#/components/schemas/BookOut"
          }
        },
        "type": "object",
        "required": [
          "today",
          "shop",
          "book"
        ],
        "title": "ShopBookOut",
        "description": "The shopkeeper's book. Also written to contract/shop.json by `make db`."
      },
      "ShopOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "locality": {
            "type": "string",
            "title": "Locality"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "locality"
        ],
        "title": "ShopOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WaitingOut": {
        "properties": {
          "scan_id": {
            "type": "string",
            "title": "Scan Id"
          },
          "customer_id": {
            "type": "string",
            "title": "Customer Id"
          },
          "display_name": {
            "type": "string",
            "title": "Display Name"
          },
          "tag": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tag"
          },
          "scanned_at": {
            "type": "string",
            "format": "date-time",
            "title": "Scanned At"
          },
          "waited_s": {
            "type": "integer",
            "title": "Waited S"
          },
          "first_time": {
            "type": "boolean",
            "title": "First Time"
          }
        },
        "type": "object",
        "required": [
          "scan_id",
          "customer_id",
          "display_name",
          "tag",
          "scanned_at",
          "waited_s",
          "first_time"
        ],
        "title": "WaitingOut"
      }
    }
  }
}